+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # API CHANGELOG
 
+#### V1.0.0b1.5.15
+9.21.19
+- Add DateField migration for CoreSubsidyRecord, LisPenden, HPDRegistration, Foreclosure, PSPreForeclosure, PSForeclosure
+
+#### V1.0.0b1.5.14
+9.21.19
+- Add DateField migration for HPDComplaint and HPDProblem
+
 #### V1.0.0b1.5.14
 9.21.19
 - Add DateField migration for HPDComplaint and HPDProblem